third_party/WebKit/LayoutTests/inspector/sources/debugger-step/debugger-step-out-document-write.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>

<script>
    document.writeln("<script>\nfunction f1() {\n  window.foo1 = true;\n}\nf1();\n <" + "/script>");
</script>
<script>
    console.assert(window.foo1, "FAIL: foo1");
</script>
<script>
function global1()
{
    debugger;
    document.writeln("<script>\nfunction f2() {\n  window.foo2 = true;\n}\nf2();\n <" + "/script>");
}
var x = global1();
</script>
<script>
    console.assert(window.foo2, "FAIL: foo2");
</script>

<script>
function test()
{
    var numberOfStepOut = 5;

    SourcesTestRunner.startDebuggerTest(step1, true);

    function step1()
    {
        SourcesTestRunner.showScriptSource("debugger-step-out-document-write.html", step2);
    }

    function step2(sourceFrame)
    {
        TestRunner.addResult("Script source was shown.");
        SourcesTestRunner.setBreakpoint(sourceFrame, 6, "", true);
        SourcesTestRunner.waitUntilPaused(step3);
        TestRunner.reloadPage(completeTest);
    }

    function step3()
    {
        var actions = ["Print"]; // First pause on breakpoint.
        for (var i = 0; i < numberOfStepOut; ++i)
            actions.push("StepOut", "Print");
        actions.push("Resume");
        SourcesTestRunner.waitUntilPausedAndPerformSteppingActions(actions);
    }

    function completeTest()
    {
        SourcesTestRunner.completeDebuggerTest();
    }
}
</script>
</head>

<body onload="runTest()">
<p>
Tests that debugger StepOut will skip inlined scripts created by document.write().
</p>
</body>
</html>
