third_party/WebKit/LayoutTests/inspector/sources/debugger-step/debugger-step-out-document-write-expected.txt

Tests that debugger StepOut will skip inlined scripts created by document.write().

Script source was shown.
Call stack:
    0)  (debugger-step-out-document-write.html:7)

Executing StepOut...
Call stack:
    0)  (debugger-step-out-document-write.html:9)

Executing StepOut...
Call stack:
    0)  (debugger-step-out-document-write.html:12)

Executing StepOut...
Call stack:
    0) global1 (debugger-step-out-document-write.html:15)
    1)  (debugger-step-out-document-write.html:18)

Executing StepOut...
Call stack:
    0)  (debugger-step-out-document-write.html:18)

Executing StepOut...
Call stack:
    0)  (debugger-step-out-document-write.html:20)

Executing Resume...
Page reloaded.
